Forest growth function + wood from the field
Improved on forest growth function yet again. Required other GIS files and some changes to the code. Growth is more realistic now, though still not very well parametrized. 
In addition, when agents now harvest a field for agriculture, they also get the wood that is on it "for free". The plot then remains cleared of forest by means of the 'wood?' switch.